More of the algorithm filled in from top down - I have implemented the iterator over all the individual inductive rules, and now need to implement the function that computes the new models generated given an individual rule.
I've also changed the signature of the Interpretation module - now it
simply exposes one function which returns the computed interpretation
when given a set of inductive rules and a candidate model for the
distinguised predicate. It seemed that this would afford the most
felxibility in implementing the fixed point calculation, and also makes
sense since the rest of the code only needs access to the resulting
interpretation, and is not performing any further processing on it.